feat: apply reviews
